requirements.txt
numpy==2.1.3
pygame==2.6.1
pyserial==3.5
RPi.GPIO==0.7.1
